--- src/components/interface/BorrowTable/components/BorrowTableRow/components/index.ts
@@ -1,4 +1,0 @@
-export * from './BorrowFixedAPR/BorrowFixedAPR';
-export * from './BorrowMaturity/BorrowMaturity';
-export * from './BorrowVariableAPY/BorrowVariableAPY';
-export * from './Debt/Debt';
