--- src/components/interface/BorrowTable/components/index.ts
@@ -1,1 +1,0 @@
-export * from './BorrowTableHead/BorrowTableHead';
