--- src/components/interface/BorrowTable/utilities/index.ts
@@ -1,1 +1,0 @@
-export { default as mapAmmToAmmTableDatum } from './mapAmmToAmmTableDatum';
