+++ src/routes/FixedBorrower/ConnectedBorrowPositionTable/BorrowTable/components/BorrowTableRow/components/BorrowFixedAPR/BorrowFixedAPR.tsx
@@ -1,8 +1,8 @@
 import TableCell from '@mui/material/TableCell';
 import React from 'react';
 
-import { formatNumber } from '../../../../../../../utilities/number';
-import { Typography } from '../../../../../../atomic/Typography/Typography';
+import { Typography } from '../../../../../../../../components/atomic/Typography/Typography';
+import { formatNumber } from '../../../../../../../../utilities/number';
 
 export type BorrowBorrowFixedAPR = {
   loading: boolean;
