+++ src/routes/FixedBorrower/ConnectedBorrowPositionTable/BorrowTable/components/BorrowTableRow/components/Debt/Debt.tsx
@@ -2,8 +2,8 @@ import Box from '@mui/material/Box';
 import TableCell from '@mui/material/TableCell';
 import React from 'react';
 
-import { formatNumber } from '../../../../../../../utilities/number';
-import { Typography } from '../../../../../../atomic/Typography/Typography';
+import { Typography } from '../../../../../../../../components/atomic/Typography/Typography';
+import { formatNumber } from '../../../../../../../../utilities/number';
 
 export type DebtProps = {
   debtInUSD: number | null | void;
